requirements.txt
llama-index-embeddings-huggingface==0.4.0
llama-index-llms-ollama
llama-index-vector-stores-typesense==0.3.0
llama-index-readers-file